requirements.txt
ansible==11.3.0
ansible-lint==25.1.3
pyOpenSSL==25.0.0
jmespath==1.0.1
